refactor(webpack-dev-server): separate createSchema from webpack-dev-server.js
